+++ src/service/AdminService.java
@@ -59,20 +59,20 @@ public interface AdminService {
     void cancelOrder(long orderId);
 
     // 통계 기능
-    int getTotalSales();
+    long getTotalSales();
 
-    Map<String, Integer> getSalesByCategory();
+    Map<String, Long> getSalesByCategory();
 
     List<String> getTopSellingMenus();
 
-    Map<String, Integer> getDailySales();
+    Map<String, Long> getDailySales();
 
-    Map<String, Integer> getSalesByPeriod(String format);
+    Map<String, Long> getSalesByPeriod(String format);
 
     // 260313 [feat]: 매출 통계 고도화 신규 메서드
     Map<String, Object> getSalesStatsByPeriod(String startDate, String endDate);
-    Map<Integer, Integer> getHourlySales();
-    Map<String, Integer> getDayOfWeekSales();
+    Map<Integer, Long> getHourlySales();
+    Map<String, Long> getDayOfWeekSales();
     List<Map<String, Object>> getTopSpenders(int limit);
 
     void exportStatisticsToCSV();
